Refactor drug price updates
Old drug price update would cause a conflict error for new drugs if it
took over 10 secs to enter the transaction for a new drug.  This was
because a new drug would not have a price and wouldn’t get a price
until the first transaction.  But if a person entering a transaction
was still entering it after the 10 sec delay before the drug price was
set automatically (which would then update all transactions including
the current one), we would get a version conflict.